Split tests between sync httpx calls and non httpx related tests
